backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.26
django-cors-headers==4.4.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
pkg_resources==0.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.13.2
tzdata==2025.2
